fix(engines): adapt to updated upstream crate APIs
Strict engine: ColumnType::Decimal is now a struct variant
(Decimal { precision, scale }) and requires an exhaustive match arm.
version_column_counts keys are u16; cast u32 schema.version when
inserting. tuple_version lookups cast to u16 consistently.

Columnar store: segment IDs are now u64 in the mutation API;
cast u32 segment_id values at all call sites. on_memtable_flushed now
returns a Result and is propagated. write_segment and the compaction
path accept additional Option parameters.

Spatial index: checkpoint_to_bytes and from_checkpoint now take an
Option<_> hints argument.

Array segments/retention: writer.finish() now takes an Option argument.

FTS manager: add a wildcard arm to the query mode match so new variants
do not cause a compile error.

Catalog: CollectionInfo gains primary and vector_primary fields;
populate them at every construction site. SqlCatalog::collection_info
gains a DatabaseId parameter. Decimal pattern updated to Decimal{..}.
column_type_to_sql adds a catch-all bytes arm.

DDL KV: add a catch-all arm to the KV type match.

+++ nodedb-lite/src/engine/strict/crud.rs
@@ -197,7 +197,7 @@ impl<S: StorageEngine> StrictEngine<S> {
                     // pad with Null for columns added after that version.
                     let old_col_count = state
                         .version_column_counts
-                        .get(&tuple_version)
+                        .get(&(tuple_version as u16))
                         .copied()
                         .unwrap_or(state.schema.columns.len());
